Allow user to swap local and remote tasks via config
