+++ app/src/main/java/edu/stanford/cardinalkit/common/Config.kt
@@ -4,4 +4,6 @@ object Config {
 
     const val TASKS_FILE = "tasks.json"
     const val CONTACTS_FILE = "contacts.json"
+    const val REMOTE_TASKS = false
+
 }
