+++ app/src/main/java/edu/stanford/cardinalkit/common/Constants.kt
@@ -43,7 +43,6 @@ object Constants {
     const val SURVEYS_USE_CASES = "surveysUseCases"
 
     const val TASKS_REPOSITORY = "tasksRepository"
-    const val TASKS_LOCAL_REPOSITORY = "tasksLocalRepository"
     const val TASKS_USE_CASES = "tasksUseCases"
 
     const val TASKLOG_REPOSITORY = "taskLogRepository"
